Move apis to root and add vm_folder support in machineConfig
1) Move api definition to root folder apis
2) Environment like VMC need to specifthe VM folder, so
add support in ProviderMachineConfig
