+++ pkg/apis/vsphereproviderconfig/register.go
