+++ pkg/cloud/vsphere/constants/constants.go
@@ -14,6 +14,7 @@ const (
 	DeleteEventAction                = "Delete"
 	ProviderDatacenter               = "datacenter"
 	ProviderDatastore                = "datastore"
+	ProviderVmFolder                 = "vm_folder"
 	ProviderResPool                  = "resource_pool"
 	ProviderNetwork                  = "network"
 	ProviderTemplate                 = "vm_template"
